Just move value_counts, isin requires dtype infra
